requirements.txt
requests
python-binance
ta
pandas
